+++ client/selfoss-sw-offline.js
@@ -7,8 +7,6 @@ async function install() {
 
     const entriesToCache =
         manifest
-            // These make no sense offline.
-            .filter((entry) => !entry.match(/^(opml)\b/))
             // We need to pass index.html through PHP to perform templating.
             .map((entry) => entry === 'index.html' ? './' : entry);